requirements.txt
memphis-py
python-dotenv
supabase-py